Fixed a permissions bug and improved usage of Net::HTTP::Persistent
File.open in its 3-argument version, takes the permissions argument
in octal, not decimal.  The constructor for Net::HTTP::persistent was
throwing an error causing issuesync to stop at that point (so it wasn't
pulling the issues). Named arguments are now being used instead, which
is a more explicit form of calling it, and at the same time the error
goes away.